Fix: return cart quantity as int always

See merge request guapa-company/backend-php!395

+++ app/Services/V3_1/CartService.php
@@ -181,7 +181,7 @@ class CartService
     }
 
     private function getCartTotalQuantity(User $user){
-        return Cart::where('user_id', $user->id)->sum('quantity');
+        return (int) Cart::where('user_id', $user->id)->sum('quantity');
     }
 
     private function calculateProductFees($product, $finalPrice): float|int
